Linux : Fix compiler issues

+++ ApplicationCode/ProjectDataModel/RimView.cpp
@@ -740,7 +740,7 @@ void RimView::addDynamicWellPathsToModel(cvf::ModelBasicList* wellPathModelBasic
 
     QDateTime currentTimeStamp;
     std::vector<QDateTime> timeStamps = ownerCase()->timeStepDates();
-    if (currentTimeStep() < timeStamps.size())
+    if (currentTimeStep() < static_cast<int>(timeStamps.size()))
     {
         currentTimeStamp = timeStamps[currentTimeStep()];
     }
